docs: add proper swagger documenations, in dto, controllers, along with api tags
