+++ api/src/app.controller.ts
@@ -2,7 +2,7 @@ import { Controller, Get } from '@nestjs/common';
 import { ApiTags, ApiOperation, ApiResponse } from '@nestjs/swagger';
 import { AppService } from './app.service';
 
-@ApiTags('health')
+@ApiTags('health (Health check endpoints)')
 @Controller()
 export class AppController {
   constructor(private readonly appService: AppService) {}
