+++ api/src/common/index.ts
@@ -1,1 +1,2 @@
 export * from './exception.filter';
+export * from './logging.interceptor';
